add commonTypes to customer
