+++ src/clients/customer/get/index.ts
@@ -1,6 +1,6 @@
 import { RestClient } from '@utils/restClient';
 import type { CustomerCardConfigs } from './types';
-import { CustomerCardResponse } from '../create/types';
+import { CustomerCardResponse } from '../commonTypes';
 
 export default function get({ customerId, cardId, config }: CustomerCardConfigs): Promise<CustomerCardResponse> {
 	return RestClient.fetch<CustomerCardResponse>(
