+++ src/clients/customer/list/index.ts
@@ -1,6 +1,6 @@
 import { RestClient } from '@utils/restClient';
 import type { CustomerCardList } from './types';
-import { CustomerCardResponse } from '../create/types';
+import { CustomerCardResponse } from '../commonTypes';
 
 export default function listAll({ customerId, config }: CustomerCardList): Promise<CustomerCardResponse[]> {
 	return RestClient.fetch<CustomerCardResponse[]>(
